requirements.txt
python-telegram-bot==20.3
langchain
openai
faiss-cpu
python-dotenv
tiktoken
